[step3] refactor: 서비스 메서드 분리
