+++ src/main/kotlin/racingcar/RacingCarApplication.kt
@@ -1,11 +1,13 @@
 package racingcar
 
 import racingcar.controller.RacingCarController
+import racingcar.service.RacingCarService
 
 object RacingCarApplication {
     @JvmStatic
     fun main(args: Array<String>) {
-        val racingCarController = RacingCarController()
+        val racingCarService = RacingCarService()
+        val racingCarController = RacingCarController(racingCarService)
         racingCarController.start()
     }
 }
